# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Freightrek Server API",
    "version": "1.0.0",
    "description": "REST API for Freightrek freight management system with admin authentication and RBAC",
    "contact": {
      "name": "Freightrek Team",
      "email": "support@freightrek.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "https://freightrekapi.vercel.app",
      "description": "Production (Vercel)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development (local)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "Hub": {
        "type": "object",
        "required": [
          "name",
          "location"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Mumbai Hub"
          },
          "location": {
            "type": "string",
            "example": "Mumbai, Maharashtra"
          },
          "capacity": {
            "type": "number",
            "example": 1000
          },
          "status": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/hub/login": {
      "post": {
        "summary": "Hub direct login",
        "tags": [
          "Hub Management"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hub login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/admin/hub/unified-login": {
      "post": {
        "summary": "Unified login for both Hub admin and Hub staff",
        "tags": [
          "Hub Management"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful (returns loginType 'hub' or 'hub_staff')"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/admin/hub": {
      "post": {
        "summary": "Create a new freight hub",
        "tags": [
          "Hub Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Hub"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hub created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error creating hub"
          }
        }
      },
      "get": {
        "summary": "Get all hubs",
        "tags": [
          "Hub Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all hubs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Hub"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/hub/{id}": {
      "get": {
        "summary": "Get hub by ID",
        "tags": [
          "Hub Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Hub ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Hub details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Hub"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Hub not found"
          }
        }
      },
      "put": {
        "summary": "Update hub",
        "tags": [
          "Hub Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Hub"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hub updated successfully"
          },
          "404": {
            "description": "Hub not found"
          }
        }
      },
      "delete": {
        "summary": "Delete hub (staff are auto-reassigned to another hub)",
        "description": "Any hub staff assigned to this hub are moved to another active hub before the hub is removed. The target hub is picked automatically (same city, then same state, then any active hub) unless reassignHubId is supplied. Hub-scoped roles are remapped to the same role name in the target hub, or cleared if it has none. The delete is rejected when staff exist and no other active hub is available.\n",
        "tags": [
          "Hub Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reassignHubId",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Hub to move this hub's staff to. Auto-selected when omitted."
          }
        ],
        "responses": {
          "200": {
            "description": "Hub deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reassignedStaffCount": {
                          "type": "number"
                        },
                        "reassignedToHub": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No active hub available for the staff, or invalid reassignHubId"
          },
          "404": {
            "description": "Hub not found"
          }
        }
      }
    },
...
    "/admin/route": {
      "post": {
        "summary": "Create a new route",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "routeName",
                  "from",
                  "to"
                ],
                "properties": {
                  "routeName": {
                    "type": "string",
                    "example": "Chennai - Bangalore Express"
                  },
                  "from": {
                    "type": "string",
                    "example": "Chennai"
                  },
                  "to": {
                    "type": "string",
                    "example": "Bangalore"
                  },
                  "branches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "Guindy",
                      "Tambaram"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Active",
                      "Inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route created successfully"
          },
          "400": {
            "description": "Validation error or route already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      },
      "get": {
        "summary": "Get all routes with pagination and filters",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "Active",
                "Inactive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of routes"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      }
    },
    "/admin/route/{id}": {
      "get": {
        "summary": "Get route by ID",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route details"
          },
          "404": {
            "description": "Route not found"
          }
        }
      },
      "put": {
        "summary": "Update route",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "routeName": {
                    "type": "string"
                  },
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  },
                  "branches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Active",
                      "Inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Route updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Route not found"
          }
        }
      },
      "delete": {
        "summary": "Delete route",
        "tags": [
          "Route Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route deleted successfully"
          },
          "404": {
            "description": "Route not found"
          }
        }
      }
    },
...
  },
...
}